test(RoomescapeControllerTest): 기존의 테스트들을 MockMvc 를 사용하도록 수정
