+++ src/test/java/roomescape/controller/RoomescapeControllerTest.java
@@ -1,16 +1,37 @@
 package roomescape.controller;
 
-import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
+import io.restassured.module.mockmvc.RestAssuredMockMvc;
+import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import static org.hamcrest.Matchers.is;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
+import static org.mockito.BDDMockito.given;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
+import roomescape.domain.Reservation;
+import roomescape.repository.ReservationRepository;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class RoomescapeControllerTest {
+@WebMvcTest(RoomescapeController.class)
+public class RoomescapeControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockitoBean
+    private ReservationRepository reservationRepository;
+
+    @BeforeEach
+    void setUp() {
+        RestAssuredMockMvc.mockMvc(mockMvc);
+    }
+
     @Test
     void 날짜가_형식에_맞지_않는_경우_400을_반환한다() {
         Map<String, String> params = new HashMap<>();
@@ -18,12 +39,36 @@ class RoomescapeControllerTest {
         params.put("date", "2023-08-053");
         params.put("time", "15:40");
 
-        RestAssured.given().log().all()
+        RestAssuredMockMvc.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
                 .when().post("/reservations")
                 .then().log().all()
                 .statusCode(400);
+    }
+
+    @Test
+    void 모든_예약_정보를_정상적으로_반환한다() {
+        // given
+        final Reservation 윌슨 = new Reservation("윌슨", LocalDate.now(), LocalTime.now());
+        final Reservation 히로 = new Reservation("히로", LocalDate.now(), LocalTime.now());
+        윌슨.setId(1L);
+        히로.setId(2L);
+        List<Reservation> reservations = List.of(
+                윌슨, 히로
+        );
+
+        given(reservationRepository.findAll()).willReturn(reservations);
+
+        RestAssuredMockMvc.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(2))
+                .body("[0].id", is(1))
+                .body("[0].name", is("윌슨"))
+                .body("[1].id", is(2))
+                .body("[1].name", is("히로"));
     }
 
     @Test
@@ -33,7 +78,7 @@ class RoomescapeControllerTest {
         params.put("date", "2023-08-05");
         params.put("time", "15:40");
 
-        RestAssured.given().log().all()
+        RestAssuredMockMvc.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
                 .when().post("/reservations")
@@ -48,7 +93,7 @@ class RoomescapeControllerTest {
         params.put("date", "2023-08-05");
         params.put("time", "15 40");
 
-        RestAssured.given().log().all()
+        RestAssuredMockMvc.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
                 .when().post("/reservations")
@@ -58,10 +103,10 @@ class RoomescapeControllerTest {
 
     @Test
     void id_값이_문자열_이므로_400을_반환한다() {
-        RestAssured.given().log().all()
+        RestAssuredMockMvc.given().log().all()
                 .pathParam("id", "일")
                 .when().delete("/reservations/{id}")
                 .then().log().all()
-                .statusCode(500);
+                .statusCode(400);
     }
 }
